refactor(abw): decompose editorial augmentation stage
