+++ apps/ai-blog-writer/apps/backend/tests/test_youtube2blog_editorial_stage.py
@@ -1,10 +1,25 @@
 import importlib
 
-from app.features.youtube2blog.config import Y2B_EDITORIAL_AUGMENTATION_MAX_OUTPUT_TOKENS
+from app.features.youtube2blog.config import (
+    Y2B_EDITORIAL_AUGMENTATION_MAX_OUTPUT_TOKENS,
+)
+from app.features.youtube2blog.content.editorial_blocks import (
+    ensure_editorial_component_boxes,
+)
+from app.features.youtube2blog.stages.editorial_augmentation_prompts import (
+    build_editorial_augmentation_prompt,
+)
+from app.features.youtube2blog.stages.editorial_augmentation_validation import (
+    sanitize_editorial_augmentation,
+)
+from app.shared.prompts import ANTI_AI_TELLS_FULL
 from shared import Stage3Output
 
 editorial_stage = importlib.import_module(
     "app.features.youtube2blog.stages.stage_editorial_augmentation"
+)
+editorial_llm = importlib.import_module(
+    "app.features.youtube2blog.stages.editorial_augmentation_llm"
 )
 
 
@@ -25,6 +40,11 @@ def _sample_stage3_output() -> Stage3Output:
 def test_editorial_augmentation_applies_components_and_markers(monkeypatch):
     stage3 = _sample_stage3_output()
 
+    monkeypatch.setattr(
+        editorial_stage,
+        "enforce_editorial_anti_ai_tells",
+        lambda content, **_kwargs: content,
+    )
     monkeypatch.setattr(
         editorial_stage,
         "_invoke_json_llm",
@@ -63,6 +83,87 @@ def test_editorial_augmentation_applies_components_and_markers(monkeypatch):
     assert output.error is None
 
 
+def test_editorial_prompt_bounds_source_fields_and_appends_guidance():
+    prompt = build_editorial_augmentation_prompt(
+        article_title=f"{'T' * 500}EXCLUDED",
+        article_content=f"{'C' * 20_000}EXCLUDED",
+        article_type="How-to Guides",
+        tone_guidance="  Keep the tone direct.  ",
+    )
+
+    assert "T" * 500 in prompt
+    assert "C" * 20_000 in prompt
+    assert "EXCLUDED" not in prompt
+    assert "Keep the tone direct." in prompt
+    assert prompt.endswith(ANTI_AI_TELLS_FULL)
+
+
+def test_editorial_markers_wrap_existing_box_without_appending_metadata_only_box():
+    content = (
+        "## Overview\n\n"
+        "> [!EDITORIAL-BOX|highlight_callout]\n"
+        "> Dense ideas become easier to scan."
+    )
+
+    updated = ensure_editorial_component_boxes(
+        content,
+        [
+            {
+                "component": "highlight_callout",
+                "justification": "Improves pacing.",
+                "placement": "After overview.",
+            }
+        ],
+    )
+
+    assert updated.count("[!EDITORIAL-BLOCK-START|highlight_callout]") == 1
+    assert updated.count("[!EDITORIAL-BLOCK-END|highlight_callout]") == 1
+    assert "[!EDITORIAL-BLOCK-LABEL|Highlight Callout]" in updated
+    assert "**Component:** Highlight Callout" in updated
+    assert "**Placement:**" not in updated
+
+
+def test_editorial_validation_normalizes_and_preserves_longer_fallback():
+    fallback = (
+        "## Overview\n\n" + " ".join(f"original-{index}" for index in range(80)) + "."
+    )
+
+    editorial = sanitize_editorial_augmentation(
+        {
+            "augmented_content": "Short.",
+            "components_added": [
+                {
+                    "component": "faq",
+                    "justification": 42,
+                    "placement": None,
+                },
+                {"component": "unsupported"},
+            ],
+            "diagnostic": {
+                "cognitive_load": "high_risk",
+                "narrative_density": "unknown",
+            },
+        },
+        fallback_content=fallback,
+    )
+
+    assert fallback in editorial["augmented_content"]
+    assert editorial["components_added"] == [
+        {
+            "component": "faq_block",
+            "justification": "42",
+            "placement": "",
+        }
+    ]
+    assert editorial["diagnostic"] == {
+        "cognitive_load": "weak",
+        "narrative_density": "strong",
+        "emphasis_clarity": "strong",
+        "reading_behavior_risk": "strong",
+    }
+    assert "[!EDITORIAL-BLOCK-START|faq_block]" in editorial["augmented_content"]
+
+
 def test_editorial_augmentation_falls_back_on_error(monkeypatch):
     stage3 = _sample_stage3_output()
 
@@ -91,9 +192,9 @@ def test_editorial_augmentation_json_llm_uses_longform_output_cap(monkeypatch):
         captured.update(kwargs)
         return StubLLM()
 
-    monkeypatch.setattr(editorial_stage, "get_vertex_llm", fake_get_vertex_llm)
+    monkeypatch.setattr(editorial_llm, "get_vertex_llm", fake_get_vertex_llm)
 
-    parsed, _raw = editorial_stage._invoke_json_llm(
+    parsed, _raw = editorial_llm.invoke_json_llm(
         prompt="Return augmented article JSON.",
         model_name="editorial-model",
     )
@@ -101,3 +202,62 @@ def test_editorial_augmentation_json_llm_uses_longform_output_cap(monkeypatch):
     assert parsed["augmented_content"] == "Body"
     assert captured["model_name"] == "editorial-model"
     assert captured["max_tokens"] == Y2B_EDITORIAL_AUGMENTATION_MAX_OUTPUT_TOKENS
+
+
+def test_editorial_augmentation_json_llm_retries_invalid_json(monkeypatch):
+    prompts: list[str] = []
+    responses = iter(
+        [
+            "not json",
+            '{"augmented_content":"Body","components_added":[]}',
+        ]
+    )
+
+    class StubLLM:
+        def invoke(self, prompt: str) -> str:
+            prompts.append(prompt)
+            return next(responses)
+
+    monkeypatch.setattr(editorial_llm, "get_vertex_llm", lambda **_kwargs: StubLLM())
+
+    parsed, raw = editorial_llm.invoke_json_llm(prompt="Original prompt")
+
+    assert parsed["augmented_content"] == "Body"
+    assert raw.startswith('{"augmented_content"')
+    assert len(prompts) == 2
+    assert "Previous invalid output:\nnot json" in prompts[1]
+
+
+def test_editorial_augmentation_uses_writing_model_for_stage_execution(monkeypatch):
+    captured: dict[str, str] = {}
+
+    def fake_invoke_json_llm(*, prompt: str, model_name: str):
+        captured["prompt"] = prompt
+        captured["model_name"] = model_name
+        return (
+            {
+                "augmented_content": "## Overview\n\nOriginal article body.",
+                "components_added": [],
+                "diagnostic": {},
+                "augmentation_summary": "",
+            },
+            "{}",
+        )
+
+    monkeypatch.setattr(editorial_stage, "_invoke_json_llm", fake_invoke_json_llm)
+    monkeypatch.setattr(
+        editorial_stage,
+        "enforce_editorial_anti_ai_tells",
+        lambda content, **_kwargs: content,
+    )
+
+    output = editorial_stage.stage_editorial_augmentation(
+        _sample_stage3_output(),
+        model_name="base-model",
+        writing_model="editorial-writing-model",
+        tone_guidance="Keep it concise.",
+    )
+
+    assert captured["model_name"] == "editorial-writing-model"
+    assert "Keep it concise." in captured["prompt"]
+    assert output.error is None
